Move module `mp_int` into own file
